+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 
 - [command_toolbox.apply_velocity] feature: new module to provide `/apply-velocity` command.
 - [command_toolbox.{burn/freeze}] feature: `/burn` and `/freeze` commands now support to select a collection of entities as target.
+- [home] feature: new `/home rename <old-name> <new-name>` command.
+